readme, error handling, colours

+++ upload.sh
@@ -9,26 +9,35 @@ function upload() {
   if [ $# -eq 2 ];
   then
 
-  notes_up_folders=("articles" "books" "coding" "discussion_stuff" "exam_solutions" "extra_things" "labs" "practice_solutions" "quiz_solutions" "study_guides" "week_notes")
-  # a vector of the target folders...
-  upload_base="./notes_upload"
-  is_inside="false"
+    if [ ! -d ${2} ] && [ ! -f ${2} ]; then
 
+      echo "need valid directory or file path...typo?"
 
-  for i in "${notes_up_folders[@]}" #note: the '#' shows no. of elements
-  do
-  	if [ $1 = $i ]; then
+    else
 
-      cp $2 ${upload_base}/${i}
-      echo "uploaded the file! check ./notes_upload/${i} for the file..."
-      is_inside="true"
-    fi
+      notes_up_folders=("articles" "books" "coding" "discussion_stuff" "exam_solutions" "extra_things" "labs" "practice_solutions" "quiz_solutions" "study_guides" "week_notes")
+      # a vector of the target folders...
+      upload_base="./notes_upload"
+      is_inside="false"
 
-  done
+      for i in "${notes_up_folders[@]}" #note: the '#' shows no. of elements
+      do
+      	if [ $1 = $i ]; then
 
-    if [ ${is_inside} = "false" ]; then
+            cp $2 ${upload_base}/${i}
 
-      echo "can't upload to '${1}'...use names exactly in ./notes_upload...typo maybe?"
+            echo "uploaded the file! check ./notes_upload/${i} for the file..."
+            is_inside="true"
+
+        fi
+
+      done
+
+      if [ ${is_inside} = "false" ]; then
+
+        echo "can't upload to '${1}'...use names exactly in ./notes_upload...typo maybe?"
+
+      fi
 
     fi
 
@@ -36,6 +45,6 @@ function upload() {
 
     echo "need topic as well as the path to the file, 2 arguments..."
 
-
   fi
+
 }
